Remove deprecated LocalMetadataServer
